# worker/docs/swagger.yaml
basePath: /
definitions:
  github_com_doniyusdinar_config-management_pkg_models.Agent:
    properties:
      id:
        type: string
      last_poll:
        type: string
      metadata:
        type: string
      registered_at:
        type: string
    type: object
  github_com_doniyusdinar_config-management_pkg_models.ConfigResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.WorkerConfig'
      poll_interval_seconds:
        type: integer
      version:
        type: integer
    type: object
  github_com_doniyusdinar_config-management_pkg_models.RegisterRequest:
    properties:
      hostname:
        type: string
      metadata:
        type: string
    type: object
  github_com_doniyusdinar_config-management_pkg_models.RegisterResponse:
    properties:
      agent_id:
        type: string
      poll_interval_seconds:
        type: integer
      poll_url:
        type: string
    type: object
  github_com_doniyusdinar_config-management_pkg_models.WorkerConfig:
    properties:
      url:
        type: string
    required:
    - url
    type: object
host: localhost:8082
info:
  contact:
    email: support@example.com
    name: API Support
  description: Worker service that executes tasks based on received configuration
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Configuration Management Worker API
  version: "1.0"
paths:
  /api/v1/agents:
    get:
      description: Get a list of all registered agents (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.Agent'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth: []
      summary: Get all registered agents
      tags:
      - agents
  /api/v1/config:
    get:
      description: Get the current active configuration for agents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.ConfigResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth: []
      summary: Get current configuration
      tags:
      - config
    post:
      consumes:
      - application/json
      description: Update the global configuration (admin only)
      parameters:
      - description: New configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.WorkerConfig'
      - default: 30
        description: Poll interval in seconds
        in: query
        name: poll_interval
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth: []
      summary: Update configuration
      tags:
      - config
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Register a new agent with the controller
      parameters:
      - description: Agent registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.RegisterResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth: []
      summary: Register a new agent
      tags:
      - agents
  /config:
    post:
      consumes:
      - application/json
      description: Receive configuration update from agent
      parameters:
      - description: Worker configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_doniyusdinar_config-management_pkg_models.WorkerConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update worker configuration
      tags:
      - config
  /health:
    get:
      description: Check if the worker service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - health
  /hit:
    get:
      description: Execute HTTP GET request to configured URL and return response
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: Response from configured URL
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute configured task
      tags:
      - task
schemes:
- http
- https
swagger: "2.0"